requirements-base.txt
# Shared production dependencies (API + worker images).
# python-libtorrent is worker-only — see requirements-worker.txt

# API process
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Auth
passlib[bcrypt]==1.7.4
# passlib 1.7.4 is incompatible with bcrypt 4.1+ (removed __about__; breaks hash + self-tests)
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9
email-validator==2.1.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Celery
celery==5.3.6
redis==5.0.4
flower==2.0.1

# HTTP / S3
httpx==0.27.0
boto3==1.34.101

# Utils
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
psutil==5.9.8
